requirements.txt
Flask==2.0.1
gunicorn==20.1.0
mysql-connector-python==8.0.25
Pillow==8.1.0
pylint==2.6.0
selenium==3.141.0
sqlparse==0.4.1
urllib3==1.26.2
w3lib==1.22.0
wasabi==0.8.0
webdriver-manager==3.4.2
